// stage-app frontend/src/app/model/FicheArrivageTransitaire.ts
export class FicheArrivageTransitaire {
  id: number;
  inventairePhysique: boolean = null;
  dateSortiePort: Date;
  verifierArticlesServTransitare0: boolean;
  verifierArticlesServTransitare1: boolean;
  verifierArticlesServTransitare2: boolean;
  user: any;
  articleFicheArrivage: any;
  images: Array<any> = [];
  imagesOutSide: Array<any> = [];
  imagesInside: Array<any> = [];
}